+++ src/components/todoFilter.tsx
@@ -1,9 +1,26 @@
 import React from 'react'
-import { useRecoilState } from 'recoil'
-import { todoListFilterState } from '../atoms/todo'
+import { selector, useRecoilState } from 'recoil'
+import { todoListFilterState, todoListState } from '../atoms/todo'
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 type ContainerProps = {}
+
+export const filteredTodoListState = selector({
+  key: 'filteredTodoListState',
+  get: ({ get }) => {
+    const filter = get(todoListFilterState)
+    const list = get(todoListState)
+
+    switch (filter) {
+      case 'Show Completed':
+        return list.filter((item) => item.isComplete)
+      case 'Show Uncompleted':
+        return list.filter((item) => !item.isComplete)
+      default:
+        return list
+    }
+  },
+})
 
 const TodoListFilterContainer: React.FC<ContainerProps> = () => {
   const [filter, setFilter] = useRecoilState(todoListFilterState)
